src/每日应用/day1-multi-agent-meeting-system/requirements.txt
openai-whisper==20231117
nltk==3.8.1
pydub==0.25.1
sqlite3